+++ src/api/auth.ts
@@ -1,22 +1,13 @@
 
 import apiInstance from './axios';
+import { AuthData, AuthProps } from '../types/auth';
 
-export interface AuthData {
-  message: string;
-  token: string;
-}
-
-export interface AuthProps {
-  email: string;
-  password: string;
-}
-
-export const login = async ({ email, password }: AuthProps) => {
+export const login = async ({ email, password }: AuthProps): Promise<AuthData> => {
   const { data } = await apiInstance.post('/users/login', { email, password })
   return data
 }
 
-export const signUp = async ({ email, password }: AuthProps) => {
+export const signUp = async ({ email, password }: AuthProps): Promise<AuthData> => {
   const { data } = await apiInstance.post('/users/create', { email, password })
   return data
 }
