+++ src/api/todo.ts
@@ -1,45 +1,27 @@
 import apiInstance from './axios'
+import { Todo, CreateTodoProps, UpdateTodoProps, DeleteTodoProps } from '../types/todo'
 
-export interface Todo {
-  title: string;
-  content: string;
-  id: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface CreateTodoProps {
-  title: string;
-  content: string;
-}
-
-export interface DeleteTodoProps {
-  todoId: string
-}
-
-export interface UpdateTodoProps extends CreateTodoProps, DeleteTodoProps{}
-
-export const getTodos = async () => {
+export const getTodos = async (): Promise<{ data: Todo[] }> => {
   const { data } = await apiInstance.get('/todos')
   return data
 }
 
-export const getTodoById = async (todoId: string) => {
+export const getTodoById = async (todoId: string): Promise<{ data: Todo }> => {
   const { data } = await apiInstance.get(`/todos/${todoId}`)
   return data
 }
 
-export const createTodo = async ({ title, content }: CreateTodoProps) => {
+export const createTodo = async ({ title, content }: CreateTodoProps): Promise<{ data: Todo }> => {
   const { data } = await apiInstance.post('/todos', { title, content })
   return data
 }
 
-export const updateTodo = async ({ todoId, title, content }: UpdateTodoProps)  => {
+export const updateTodo = async ({ todoId, title, content }: UpdateTodoProps): Promise<{ data: Todo }>  => {
   const { data } = await apiInstance.put(`/todos/${todoId}`, { title, content })
   return data
 }
 
-export const deleteTodo = async (todoId: DeleteTodoProps) => {
+export const deleteTodo = async ({ todoId }: DeleteTodoProps): Promise<{ data: null }> => {
   const { data } = await apiInstance.delete(`/todos/${todoId}`)
   return data
 }
