+++ src/types/auth.ts
@@ -1,0 +1,9 @@
+export interface AuthData {
+  message: string;
+  token: string;
+}
+
+export interface AuthProps {
+  email: string;
+  password: string;
+}
